Removed the /nodcy/ common block
Removed the "nodcy" common block. This was used to specify options for
simulating decay channels within the QGSM.

Retained option flag modeling of decay or not. This is hosted in the
data module and publicly available to change. This should perhaps be changed
so to be an option for the simulation class itself.