Adjusted "Flavor" to use if-else instead of switch-case
Execution durations seemed about the same, statistically.
Additionally, the "if-else" construct is more "fool-proof"
in terms of risk introduction (for example, could evaluate
cases after selecting one, which may change the outcome).

The example case was encountered during the initial switch to
the switch-case construct, and it seemed to read easier
with more obvious logical flow using the if-else within
this sub.